search performance and code cleanup
The search performance was pretty poor before.  This change speeds things up a little by only querying once per expectation and searching for all expected IDs based on the fetched list.
A new abstraction was introduced for dealing with Queries as an object instead of as a String to help make it easier to differentiate between queries among different sources
new metrics and tests are in progress