XML Feed Debug Display support
XML Feed Debug adds some additional information to the output so that we can look at things like vse-keys AND urls and consider both when deciding whether an expectation is in the result set.  Additional info such as base scores is also returned.  With this fetcher, the profiler results data structure can have output more data in a root cause report.  The XML Feed Debug Display has to be added to the Engine instance otherwise this fetcher will fail.